[analyzer] Introduce a new ForLoop class.
This class is part of the public API, and contains the functionality
that's common to ForElement and ForStatement.

Adding this class allows us to change the type of
`ForLoopParts.parent` from `AstNode?` to `ForLoop`. I intend to make
use of this in a follow-up CL that makes improvements to the "remove
dead code" fix, as part of my work on sound flow analysis.

Change-Id: I93ca15d2eeea495064b70690584b7a91058510f7
Bug: https://github.com/dart-lang/sdk/issues/60438
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/423560
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>