[beta] Roll dart_style 3.1.0 into SDK.
Issue description: This brings the 3.1.0 dart_style into the SDK. It has
a number of requested style fixes since 3.7 and adds support for the
3.8 language version, including support for null-aware elements.

Why cherry-pick: We want as much experience with the new style on beta
before shipping as possible. Also, this is needed to be able to format
3.8 code using null-aware elements.

Risk: Moderate. It's just a DEPS change so the cherrypick itself is
simple, but the style changes are significant. To mitigate risk, the
new style only applies to code at language version 3.8 or higher. I've
tested the new formatter by running it on a huge corpus of pub packages.

Cherry-pick: https://dart-review.googlesource.com/c/sdk/+/419990
Change-Id: I60d2d97a4c54185e33acbe57cf217178c5433e3d
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/422022
Reviewed-by: Alexander Thomas <athom@google.com>
Reviewed-by: Nate Bosch <nbosch@google.com>
Commit-Queue: Bob Nystrom <rnystrom@google.com>
Auto-Submit: Bob Nystrom <rnystrom@google.com>