Fine. Remove unexpected transitions from BaseNameMembers.
We don't have tests for these, and because of specific order we add members, we cannot test them. So, it is better to say that these transitions don't work.

If we are wrong, and it happens on any test code in the future, we
will write corresponding unit tests and so show how it works.

Change-Id: Ifdd0f258e2ddab00850782fccc6b6e3efc9f156f
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/424686
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Paul Berry <paulberry@google.com>