[cfe] Create *FragmentDeclaration
This adds *FragmentDeclaration interfaces for member declaration in order to separate the interface used by the enclosing builder (which is shared between both syntactical and synthetic declarations) and by the body builder context (which is only needed for syntactical declarations). This also prepares for implementing the getter/setter aspect of a field fragment through the GetterDeclaration/SetterDeclaration interfaces.

Change-Id: I314dfe084b2c9d48782d2f6a0c1ae20ff5f7d8c6
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/424821
Reviewed-by: Chloe Stefantsova <cstefantsova@google.com>
Commit-Queue: Johnni Winther <johnniwinther@google.com>