[analyzer] Dot shorthands: Allow potentially constant DotShorthandPropertyAccesses.
When we test for potentially-constant constants (ie. in asserts), we need to also test whether the identifier of a `DotShorthandPropertyAccess` is constant or not.

This CL adds one bit of logic and a few unit tests.
The language test `language/dot_shorthands/equality/equality_test` also covers this behavior and is passing.

Bug: https://github.com/dart-lang/sdk/issues/59835
Change-Id: I4e3daeb008c7dbadbb71d988012bf2e521df2e84
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/424661
Reviewed-by: Chloe Stefantsova <cstefantsova@google.com>
Commit-Queue: Kallen Tu <kallentu@google.com>