[dart2wasm] Move uncheckedEntry handling in updateable function wrappers to be branched at runtime.
For selectors that can be overridden in a dynamic module we produce a special dispatch wrapper that each module implements to do dispatch based on the details of its own dispatch table. Then at runtime dispatches of these selectors get sent to the wrapper for the module that owns the class of the receiver object.

uncheckedEntry references occupy a different space in the dispatch table and so require their own wrapper logic.

Currently we create separate wrappers for the unchecked and checked references. Now we use the same wrapper and have it take an extra argument that indicates if the call is checked or not. We then delegate to the appropriate offsets in the table based on this new flag.

Change-Id: I2004575d06c76c0a80570347747e179e8561463a
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/422344
Reviewed-by: Martin Kustermann <kustermann@google.com>