Allow analytics gathering for any clients that can display a message
This test was added before the unified analytics package and was
intended to guard against gathering analytics where we were not allowed
to do so. We can now gather analytics from other clients as long as we
have displayed the appropriate message, and that should be a sufficient
check at this point.

Change-Id: I1ec68439cb1048d99fbbdec2084332e8e830235a
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/417861
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Phil Quitslund <pquitslund@google.com>
Reviewed-by: Kenzie Davisson <kenzieschmoll@google.com>