[analysis_server] Make LSP executeCommand() a shared handler to support the legacy server
As a pre-cursor to supporting `executeCommand()` over DTD, this converts the handler from an LspHandler to a SharedHandler so it can work with either a base LSP or Legacy server. Not all commands themselves work in both modes, but the executeCommand() handler itself and some basic commands (like `logAction`, which I used for testing) do.

Change-Id: I90de88a1870649fb121cc9fa45732b4d03a45504
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/425263
Commit-Queue: Elliott Brooks <elliottbrooks@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Elliott Brooks <elliottbrooks@google.com>