[beta][vm] Evaluate abstract origin DIE offsets sooner in DWARF assembly output.
This is a cherry pick of the change which fixes b/410018517
(Flutter AOT snapshotting step timing out with Xcode 16.x)

Issue description : Xcode 16.x has a newer version of clang tools
and the assembler in this version was taking a long time to assemble
the output from the AOT compiler.

What is the fix : This change greatly improves assembler performance in
XCode 16, presumably by convincing the assembler it doesn't need to keep
track of relocations for abstract origin labels.

Why Cherry Pick : The cherry pick is needed as a number of iOS developers
have to switch to XCode 16 as Apple is dropping support for Xcode 15.4 and
any apps built with 15.4 can't be published to the App Store after April 24th.

Risk : low

TEST=ci

Bug: https://github.com/dart-lang/sdk/issues/43299
Bug: b/410018517
Change-Id: If866f0bca7588867e4a30512b5a331dbd46a3e0d
Cherry-pick: https://dart-review.googlesource.com/c/sdk/+/422540
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/425982
Reviewed-by: Slava Egorov <vegorov@google.com>
Commit-Queue: Siva Annamalai <asiva@google.com>