Fine. Add a few tests that ask element model manually.
These tests just record requirements, but don't verify that these
requirements cause re-computation of the result if changed. My current
thinking is that we might have two main axis of tests: library manifests
that verify that we give new IDs to changed elements; and requirements
that verify that asking a piece of data from the element model is
recorded as a requirement (or not, if this piece of data already cause
the element ID to change).

We need at least one test for each requirement kind that verifies that
such requirement does cause re-computation. But we don't need to do this
in every test. So, we avoid combinatorial explosion of cases? E.g. if
there are two requirements recorded, we don't want to try all 4
combinations how they change.

Change-Id: Iab9b0f1cd7749bc4f3c404efa340b4722bb0214e
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/426661
Reviewed-by: Paul Berry <paulberry@google.com>
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>